Fix LibreSSL returns 0 on failure
In some cases LibreSSL's s_client returns a status of 0 even when the connection fails. In the cases in which this happens there always seems to be a line in the output of the form "SSL handshake has read 0 bytes and written ... bytes". The commit fixes the problem by not treating a connection as successful if the string "SSL handshake has read 0 bytes and written" is present.