Test for conformance to NIST SP 800-52
This commit adds a --nist switch (and a corresponding NIST environment variable). If the switch is used, then testssl.sh will perform some checks for conformance to NIST Special Publication 800-52 Revision 2. For example, ratings of cipher suites will be based on whether they are permitted by SP 800-52. In addition, when cipher suites that offer forward secrecy are offered, ratings for the ephemeral ellitic curves and/or finite field groups offered are based on conformance to SP 800-56A Revision 3 (and in particular Appendix D of that document).